Add the skeleton of a game loop
This sets up a basic game loop. When all slots are filled, this will
return that the game is a draw.  If a win condition is met before that
it set the result to the winner, which will break the game loop. Each
turn we prompt the player to drop a stone in a column.